v3.0.41 jar docs jmock now available issue #142
