+++ LICENSE.md
@@ -1,0 +1,20 @@
+# Academic Use License
+
+## Permissions
+- Private use
+- Academic use only (non-commercial)
+
+## Limitations
+- Liability
+- Warranty
+
+## Conditions
+- Use is restricted to academic purposes only.
+- Copying, modifying, publishing, or distributing this software without explicit written permission from the original development team is strictly prohibited.
+- Any commercial use of this software is not permitted.
+
+This software is intended solely for academic purposes and may not be used in commercial settings. Unauthorized copying, modification, or distribution is strictly forbidden unless prior written consent is obtained from the original authors or development team.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES, OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT, OR OTHERWISE, ARISING FROM, OUT OF, OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+For more information, please contact [contact information of the development team].
