+++ src/main/java/org/project/Main.java
@@ -1,23 +1,33 @@
 package org.project;
 
+import java.util.Scanner;
 import org.project.View.CLIView;
 import org.project.View.GUIView;
 
 //TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
 // click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
 public class Main {
-    public static void main(String[] args)  {
 
+  public static void main(String[] args) {
 
-        if(args.length == 0 || args[0].equals("--gui")) {
-            GUIView.main(args);
-        }
-        else if (args[0].equals("--cli")) {
-            new CLIView().runMenu();
-        }
-        else {
-            System.out.println("Invalid command");
-        }
+    boolean answer = true;
+    Scanner scanner = new Scanner(System.in);
 
+    while (answer) {
+      System.out.print("Do you want to use GUI UML editor? (yes/y or no/n): ");
+      String input = scanner.nextLine().toLowerCase();
+
+      if (input.equals("yes") || input.equals("y")) {
+        answer = false;
+        GUIView.main(args);
+      } else if (input.equals("no") || input.equals("n")) {
+        answer = false;
+        new CLIView().runMenu();
+      } else {
+        System.out.println("Please enter a valid command.");
+      }
     }
+
+    scanner.close();
+  }
 }
