+++ src/main/java/org/project/View/GUIView.java
@@ -7,29 +7,28 @@ import javafx.scene.Scene;
 import javafx.stage.Stage;
 
 public class GUIView extends Application {
-    @Override
-    public void start(Stage stage) {
 
-        try {
+  @Override
+  public void start(Stage stage) {
 
-            FXMLLoader loader = new FXMLLoader(getClass().getResource("/Main.fxml"));
-            Parent root = loader.load();
-            Scene scene = new Scene(root);
-            stage.setScene(scene);
-            stage.setTitle("UML EDITOR");
-            stage.show();
+    try {
 
-        }catch (Exception e ){
-            e.printStackTrace();
-        }
+      FXMLLoader loader = new FXMLLoader(getClass().getResource("/Main.fxml"));
+      Parent root = loader.load();
+      Scene scene = new Scene(root);
+      stage.setScene(scene);
+      stage.setTitle("UML EDITOR");
+      stage.show();
 
-
-
-
-
+    } catch (Exception e) {
+      e.printStackTrace();
     }
-    public static void main(String[] args) {
-        launch(args);
-    }
+
+
+  }
+
+  public static void main(String[] args) {
+    launch(args);
+  }
 }
